requirements.txt
aiofiles==23.2.1
albucore==0.0.23
albumentations==2.0.4
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
asttokens==3.0.1
async-timeout==5.0.1
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.10.1
comm==0.2.3
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.19
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
elastic-transport==9.2.0
elasticsearch==9.2.0
exceptiongroup==1.3.1
executing==2.2.1
fastapi==0.121.1
ffmpy==1.0.0
filelock==3.20.0
flatbuffers==25.9.23
fonttools==4.60.1
fsspec==2025.10.0
gradio==4.44.1
gradio_client==1.3.0
h11==0.16.0
hanlp==2.1.3
hanlp_common==0.0.23
hanlp_downloader==0.0.25
hanlp_trie==0.0.5
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.11
importlib_resources==6.5.2
ipykernel==7.1.0
ipython==8.38.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.8.0
jupyter_core==5.9.1
keyboard==0.13.5
kiwisolver==1.4.9
langchain==1.2.0
langchain-core==1.2.7
langchain-huggingface==1.2.0
langgraph==1.0.7
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.7
langgraph-sdk==0.3.3
langsmith==0.6.6
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matplotlib==3.10.7
matplotlib-inline==0.2.1
mdurl==0.1.2
MouseInfo==0.1.3
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
nvidia-ml-py==13.590.48
omegaconf==2.3.0
onnx==1.14.0
onnxruntime==1.15.1
onnxruntime-gpu==1.18.0
onnxslim==0.1.31
openai==2.7.2
opencv-python==4.9.0.80
opencv-python-headless==4.11.0.86
orjson==3.11.4
ormsgpack==1.12.2
packaging==25.0
pandas==2.3.3
parso==0.8.5
phrasetree==0.0.9
pillow==10.4.0
prompt_toolkit==3.0.52
protobuf==6.33.0
psutil==5.9.8
pure_eval==0.2.3
py-cpuinfo==9.0.0
PyAutoGUI==0.9.54
pyclipper==1.4.0
pycocotools==2.0.7
pydantic==2.12.4
pydantic_core==2.41.5
pydub==0.25.1
PyGetWindow==0.0.9
Pygments==2.19.2
PyMsgBox==2.0.1
pynvml==13.0.1
pyparsing==3.2.5
pyperclip==1.11.0
pyreadline3==3.5.4
PyRect==0.2.0
PyScreeze==1.0.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-multipart==0.0.20
pytweening==1.2.0
pytz==2025.2
pywin32==311
PyYAML==6.0.1
pyzmq==27.1.0
rapidocr==3.5.0
redis==7.0.1
regex==2026.1.15
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
ruff==0.14.4
safetensors==0.4.3
scikit-learn==1.7.2
scipy==1.13.0
seaborn==0.13.2
semantic-version==2.10.0
sentence-transformers==5.2.0
sentencepiece==0.2.1
shapely==2.1.2
shellingham==1.5.4
simsimd==6.5.3
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.49.3
stringzilla==4.2.3
supervision==0.22.0
sympy==1.14.0
tenacity==9.1.2
termcolor==3.3.0
threadpoolctl==3.6.0
timm==1.0.14
tokenizers==0.22.2
tomlkit==0.12.0
toposort==1.5
torch==2.9.1
torchaudio==2.9.1
torchvision==0.24.1
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.3
typer==0.20.0
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ultralytics-thop==2.0.18
urllib3==2.5.0
uuid_utils==0.14.0
uvicorn==0.38.0
wcwidth==0.2.14
websockets==12.0
xxhash==3.6.0
zstandard==0.25.0
